support x86

+++ gles2app3/src/gles2text.cpp
@@ -115,7 +115,7 @@ struct atlas
                 rowh = 0;
             }
             roww += g->bitmap.width + 1;
-            rowh = std::max(rowh, g->bitmap.rows);
+            rowh = std::max(rowh, static_cast<int>(g->bitmap.rows));
         }
 
         w = std::max(w, roww);
@@ -171,7 +171,7 @@ struct atlas
             c[i].tx = ox / (float)w;
             c[i].ty = oy / (float)h;
 
-            rowh = std::max(rowh, g->bitmap.rows);
+            rowh = std::max(rowh, static_cast<int>(g->bitmap.rows));
             ox += g->bitmap.width + 1;
         }
